Move away from projectionist
The only use of it I still had was jumping up to parent latex file, but a new function does that now
(and better, it will always jump to the parent, no matter the input level (as long as there is only
one input of the current file, it jumps to the first match if there are multiple)).

Moved away from it in all other file types by having all docs and unit tests for specifc src files within
those files (for julia, wasn't really using it for others anyway).

This should allow for me to use some session manager again, there was some strange behaviour trying to use projectsionist with a session manager before.